Yarnspinner for Godot implemented
Hoorays and happy days, Yarnspinner now works in our project.
This now means that we can reuse some of the content and formatting we had on the previous unity project, and continue learning Yarnspinner, just in a different engine.

Props to @AdamDJames - Master Programmer, for getting me to try it on interact as opposed to in _ready. Genius Move.